examples/predictive_analytics/requirements.txt
pydantic-ai-backend[docker,console]
pydantic-ai>=0.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
